pyproject.toml
[tool.poetry]
name = "grook"
version = "0.1.0"
description = "Group Book Web App"
authors = ["Nigel Gott <nigel.gott@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
boto3 = "^1.9"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
pytest-flask = "^0.14.0"
moto = "^1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

